updated msString

+++ msString.h
@@ -1,3 +1,4 @@
+/* 07-Feb-2024: updated the return value for the msLength function to be long int */
 #ifndef __MSSTRING__
 #define __MSSTRING__
 
@@ -11,7 +12,7 @@ extern msString  msSetString     (char *);
 extern char     *msGetString     (msString);
 extern void      msCopy          (msString *, msString);
 extern void      msConcatenate   (msString *, msString);
-extern int       msLength        (msString);
+extern long int  msLength        (msString);
 extern int       msCompare       (msString, msString);
 extern int       msCompareString (msString, char *);
 static void      msError         (char *);
